refactor(dsl): move selector groups to own package
