+++ bundles/org.dataflowanalysis.analysis/src/org/dataflowanalysis/analysis/dsl/groups/DataDestinationSelectors.java
@@ -1,29 +1,27 @@
-package org.dataflowanalysis.analysis.dsl;
+package org.dataflowanalysis.analysis.dsl.groups;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.StringJoiner;
 import org.apache.log4j.Logger;
+import org.dataflowanalysis.analysis.dsl.AbstractParseable;
+import org.dataflowanalysis.analysis.dsl.AnalysisConstraint;
 import org.dataflowanalysis.analysis.dsl.context.DSLContext;
 import org.dataflowanalysis.analysis.dsl.selectors.*;
 import org.dataflowanalysis.analysis.utils.LoggerManager;
 import org.dataflowanalysis.analysis.utils.ParseResult;
 import org.dataflowanalysis.analysis.utils.StringView;
 
-/**
- * Represents the source vertex {@link AbstractSelector} matched by an {@link AnalysisConstraint}
- */
-public class VertexSourceSelectors extends AbstractParseable {
-    private static final String DSL_KEYWORD = "vertex";
-    private static final Logger logger = LoggerManager.getLogger(VertexSourceSelectors.class);
-
+public class DataDestinationSelectors extends AbstractParseable {
+    private static final Logger logger = LoggerManager.getLogger(DataDestinationSelectors.class);
+    private static final String DSL_KEYWORD = "data";
     private final List<AbstractSelector> selectors;
 
-    public VertexSourceSelectors() {
-        selectors = new ArrayList<>();
+    public DataDestinationSelectors() {
+        this.selectors = new ArrayList<>();
     }
 
-    public VertexSourceSelectors(List<AbstractSelector> selectors) {
+    public DataDestinationSelectors(List<AbstractSelector> selectors) {
         this.selectors = selectors;
     }
 
@@ -32,7 +30,7 @@ public class VertexSourceSelectors extends AbstractParseable {
     }
 
     public List<AbstractSelector> getSelectors() {
-        return new ArrayList<>(selectors);
+        return selectors;
     }
 
     @Override
@@ -49,13 +47,13 @@ public class VertexSourceSelectors extends AbstractParseable {
     }
 
     /**
-     * Parses the {@link VertexSourceSelectors} of an {@link AnalysisConstraint}.
+     * Parses the {@link DataSourceSelectors} of an {@link AnalysisConstraint}.
      * @param string String view on the string that is parsed
      * @param context DSL context used during parsing
-     * @return Returns a {@link ParseResult} that may contain the {@link VertexSourceSelectors} of the
+     * @return Returns a {@link ParseResult} that may contain the {@link DataSourceSelectors} of the
      * {@link AnalysisConstraint}
      */
-    public static ParseResult<VertexSourceSelectors> fromString(StringView string, DSLContext context) {
+    public static ParseResult<DataDestinationSelectors> fromString(StringView string, DSLContext context) {
         string.skipWhitespace();
         if (string.invalid()) {
             return ParseResult.error("Unexpected end of input!");
@@ -75,26 +73,19 @@ public class VertexSourceSelectors extends AbstractParseable {
         List<AbstractSelector> selectors = new ArrayList<>();
         while (!string.invalid()) {
             string.skipWhitespace();
-            var listSelector = VertexCharacteristicsListSelector.fromString(string, context);
+            var listSelector = DataCharacteristicListSelector.fromString(string, context);
             if (listSelector.successful()) {
                 selectors.add(listSelector.getResult());
                 continue;
             }
-
-            var selector = VertexCharacteristicsSelector.fromString(string, context);
+            var selector = DataCharacteristicsSelector.fromString(string, context);
             if (selector.successful()) {
                 selectors.add(selector.getResult());
                 continue;
             }
-
-            var typeSelector = VertexTypeSelector.fromString(string, context);
-            if (typeSelector.successful()) {
-                selectors.add(typeSelector.getResult());
-                continue;
-            }
-            var vertexNameSelector = VertexNameSelector.fromString(string, context);
-            if (vertexNameSelector.successful()) {
-                selectors.add(vertexNameSelector.getResult());
+            var nameSelector = VariableNameSelector.fromString(string, context);
+            if (nameSelector.successful()) {
+                selectors.add(nameSelector.getResult());
                 continue;
             }
             var anySelector = AnySelector.fromString(string, context);
@@ -108,7 +99,7 @@ public class VertexSourceSelectors extends AbstractParseable {
             string.setPosition(position);
             return ParseResult.error("Keyword " + DSL_KEYWORD + " is missing any selectors!");
         }
-        VertexSourceSelectors vertexSourceSelectors = new VertexSourceSelectors(selectors);
-        return ParseResult.ok(vertexSourceSelectors);
+        DataDestinationSelectors dataSourceSelectors = new DataDestinationSelectors(selectors);
+        return ParseResult.ok(dataSourceSelectors);
     }
 }
