+++ bundles/org.dataflowanalysis.analysis/src/org/dataflowanalysis/analysis/dsl/groups/VertexSourceSelectors.java
@@ -1,9 +1,11 @@
-package org.dataflowanalysis.analysis.dsl;
+package org.dataflowanalysis.analysis.dsl.groups;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.StringJoiner;
 import org.apache.log4j.Logger;
+import org.dataflowanalysis.analysis.dsl.AbstractParseable;
+import org.dataflowanalysis.analysis.dsl.AnalysisConstraint;
 import org.dataflowanalysis.analysis.dsl.context.DSLContext;
 import org.dataflowanalysis.analysis.dsl.selectors.*;
 import org.dataflowanalysis.analysis.utils.LoggerManager;
@@ -11,19 +13,19 @@ import org.dataflowanalysis.analysis.utils.ParseResult;
 import org.dataflowanalysis.analysis.utils.StringView;
 
 /**
- * Represents the source data {@link AbstractSelector} matched by an {@link AnalysisConstraint}
+ * Represents the source vertex {@link AbstractSelector} matched by an {@link AnalysisConstraint}
  */
-public class DataSourceSelectors extends AbstractParseable {
-    private static final Logger logger = LoggerManager.getLogger(DataSourceSelectors.class);
-    private static final String DSL_KEYWORD = "data";
+public class VertexSourceSelectors extends AbstractParseable {
+    private static final String DSL_KEYWORD = "vertex";
+    private static final Logger logger = LoggerManager.getLogger(VertexSourceSelectors.class);
 
     private final List<AbstractSelector> selectors;
 
-    public DataSourceSelectors() {
+    public VertexSourceSelectors() {
         selectors = new ArrayList<>();
     }
 
-    public DataSourceSelectors(List<AbstractSelector> selectors) {
+    public VertexSourceSelectors(List<AbstractSelector> selectors) {
         this.selectors = selectors;
     }
 
@@ -49,13 +51,13 @@ public class DataSourceSelectors extends AbstractParseable {
     }
 
     /**
-     * Parses the {@link DataSourceSelectors} of an {@link AnalysisConstraint}.
+     * Parses the {@link VertexSourceSelectors} of an {@link AnalysisConstraint}.
      * @param string String view on the string that is parsed
      * @param context DSL context used during parsing
-     * @return Returns a {@link ParseResult} that may contain the {@link DataSourceSelectors} of the
+     * @return Returns a {@link ParseResult} that may contain the {@link VertexSourceSelectors} of the
      * {@link AnalysisConstraint}
      */
-    public static ParseResult<DataSourceSelectors> fromString(StringView string, DSLContext context) {
+    public static ParseResult<VertexSourceSelectors> fromString(StringView string, DSLContext context) {
         string.skipWhitespace();
         if (string.invalid()) {
             return ParseResult.error("Unexpected end of input!");
@@ -75,19 +77,26 @@ public class DataSourceSelectors extends AbstractParseable {
         List<AbstractSelector> selectors = new ArrayList<>();
         while (!string.invalid()) {
             string.skipWhitespace();
-            var listSelector = DataCharacteristicListSelector.fromString(string, context);
+            var listSelector = VertexCharacteristicsListSelector.fromString(string, context);
             if (listSelector.successful()) {
                 selectors.add(listSelector.getResult());
                 continue;
             }
-            var selector = DataCharacteristicsSelector.fromString(string, context);
+
+            var selector = VertexCharacteristicsSelector.fromString(string, context);
             if (selector.successful()) {
                 selectors.add(selector.getResult());
                 continue;
             }
-            var nameSelector = VariableNameSelector.fromString(string, context);
-            if (nameSelector.successful()) {
-                selectors.add(nameSelector.getResult());
+
+            var typeSelector = VertexTypeSelector.fromString(string, context);
+            if (typeSelector.successful()) {
+                selectors.add(typeSelector.getResult());
+                continue;
+            }
+            var vertexNameSelector = VertexNameSelector.fromString(string, context);
+            if (vertexNameSelector.successful()) {
+                selectors.add(vertexNameSelector.getResult());
                 continue;
             }
             var anySelector = AnySelector.fromString(string, context);
@@ -101,7 +110,7 @@ public class DataSourceSelectors extends AbstractParseable {
             string.setPosition(position);
             return ParseResult.error("Keyword " + DSL_KEYWORD + " is missing any selectors!");
         }
-        DataSourceSelectors dataSourceSelectors = new DataSourceSelectors(selectors);
-        return ParseResult.ok(dataSourceSelectors);
+        VertexSourceSelectors vertexSourceSelectors = new VertexSourceSelectors(selectors);
+        return ParseResult.ok(vertexSourceSelectors);
     }
 }
